pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "docchat"
version = "1.0.0"
description = "RAG-powered document chatbot supporting OpenAI, Anthropic, Gemini, and local Ollama"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["rag", "llm", "chatbot", "document-ai", "langchain", "streamlit"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "streamlit>=1.30.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.2.0",
    "langchain-text-splitters>=0.2.0",
    "langchain-core>=0.2.0",
    "faiss-cpu>=1.7.4",
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "tiktoken>=0.6.0",
    "sentence-transformers>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.coverage.run]
source = ["core", "llm_config"]
omit = ["tests/*"]
